user authentication with token added

+++ controllers/expense.js
@@ -1,45 +1,54 @@
 const Expense = require('../models/Expenses');
 
+function isstringinvalid(string){
+    if(string== undefined || string.length === 0){
+        return true
+    } else{
+        return false 
+    }
+}
+
 exports.insertExpense = async (req,res,next)=>{
     try{
-        if(!req.body.expense){
-            throw new Error('Enter an amount')
+        const {expenseamount,description,category}=req.body;
+        if(isstringinvalid(expenseamount)||isstringinvalid(description)||isstringinvalid(category)){
+            return res.status(400).json({err:"Bad parameters. Something is missing"})
         }
-    const expense=req.body.expense;
-    const description=req.body.description;
-    const category=req.body.category;
 
-    const data = await Expense.create({amount: expense, description:description, category:category});
-    res.status(201).json({newExpenseDetail:data});
+        const expense = await req.user.createExpense({expenseamount, description, category}) // "or use" Expense.create({expenseamount, description, category,userId:req.user.id});
+        return res.status(201).json({expense,success:true});
     } catch(err){
-        res.status(500).json({
-            error:err
-        })
+        return res.status(500).json({success:false,error:err})
     }
 } 
 
-exports.getExpenses = async (req,res,next)=>{
+exports.getexpenses = async (req,res,next)=>{
     try{
-        const expenses=await Expense.findAll();
-        res.status(200).json({allExpenses:expenses});
-    } catch(error){
-        console.log('Get expense is failing',JSON.stringify(error));
-        res.status(500).json({error:error})
+        const expenses=await req.user.getExpenses();   //"or use"req.user.getExpenses()  Expense.findAll({where:{userId:req.user.id}});
+        return res.status(200).json({expenses,success:true});
+    } catch(err){
+        console.log('Get expense is failing',JSON.stringify(err));
+        return res.status(500).json({error:err,success:false})
     }
 }
 
 exports.deleteExpense = async (req,res,next)=>{
     try{
-        if(req.params.id=='undefined'){
-            console.log('ID is missing');
-            return res.status(400).json({err:'ID is missing'})
-        }
         const eId=req.params.id;
-        await Expense.destroy({where:{id:eId}});
-        res.sendStatus(200);
+
+        if(isstringinvalid(eId)){
+            console.log('ID is missing');
+            return res.status(400).json({success:false,message:'ID is missing'})
+        }
+        
+        const noofrows = await Expense.destroy({where:{id:eId,userId:req.user.id}});
+        if(noofrows===0){
+            return res.status(404).json({success:false,message:"Expense doesnt belong to the user"})
+        }
+        return res.status(200).json({success:true,message:"Deleted Succesfully"});
     } catch(error){
         console.log(err);
-        res.status(500).json(err);
+        return res.status(500).json({success:false,message:"Failed"});
     }
 }
